chore(react-server): rename to `server/ssr/browser` triple
