+++ examples/react-server/src/adapters/node.tsx
@@ -1,4 +1,4 @@
 import { createMiddleware } from "@hattip/adapter-node/native-fetch";
-import { handler } from "../entry-server";
+import { handler } from "../entry-ssr";
 
 export default createMiddleware((ctx) => handler(ctx.request));
